+++ app/gamehub-player/src/test/hooks/jinaga-example.test.ts
